+++ tests/dummy/app/controllers/application.js
@@ -2,11 +2,11 @@
 
 import Controller from '@ember/controller';
 
-export default Controller.extend({
-  urlsExample:
-    'Today, Twitter is updating embedded Tweets to enable a richer photo experience: https://t.co/XdXRudPXH5',
+export default class ApplicationController extends Controller {
+  urlsExample =
+    'Today, Twitter is updating embedded Tweets to enable a richer photo experience: https://t.co/XdXRudPXH5';
 
-  urlsExampleEntities: {
+  urlsExampleEntities = {
     hashtags: [],
     symbols: [],
     urls: [
@@ -15,30 +15,30 @@ export default Controller.extend({
         expanded_url:
           'https://blog.twitter.com/2013/rich-photo-experience-now-in-embedded-tweets-3',
         display_url: 'blog.twitter.com/2013/rich-phot\u2026',
-        indices: [80, 103],
-      },
+        indices: [80, 103]
+      }
     ],
-    user_mentions: [],
-  },
+    user_mentions: []
+  };
 
-  hashtagsExample: 'Loved #devnestSF',
+  hashtagsExample = 'Loved #devnestSF';
 
-  hashtagsExampleEntities: {
+  hashtagsExampleEntities = {
     hashtags: [
       {
         text: 'devnestSF',
-        indices: [6, 16],
-      },
+        indices: [6, 16]
+      }
     ],
     symbols: [],
     urls: [],
-    user_mentions: [],
-  },
+    user_mentions: []
+  };
 
-  userMentionsExample:
-    'We\u2019re excited to work closely with the external technical community and continue @twittereng\u2019s work with open source. cc @TwitterOSS',
+  userMentionsExample =
+    'We\u2019re excited to work closely with the external technical community and continue @twittereng\u2019s work with open source. cc @TwitterOSS';
 
-  userMentionsExampleEntities: {
+  userMentionsExampleEntities = {
     hashtags: [],
     symbols: [],
     urls: [],
@@ -48,21 +48,21 @@ export default Controller.extend({
         name: 'Twitter Engineering',
         id: 6844292,
         id_str: '6844292',
-        indices: [81, 92],
+        indices: [81, 92]
       },
       {
         screen_name: 'TwitterOSS',
         name: 'Twitter Open Source',
         id: 376825877,
         id_str: '376825877',
-        indices: [121, 132],
-      },
-    ],
-  },
+        indices: [121, 132]
+      }
+    ]
+  };
 
-  mediaExample: 'Four more years. http://t.co/bAJE6Vom',
+  mediaExample = 'Four more years. http://t.co/bAJE6Vom';
 
-  mediaExampleEntities: {
+  mediaExampleEntities = {
     hashtags: [],
     symbols: [],
     urls: [],
@@ -83,25 +83,25 @@ export default Controller.extend({
           medium: {
             w: 600,
             h: 399,
-            resize: 'fit',
+            resize: 'fit'
           },
           thumb: {
             w: 150,
             h: 150,
-            resize: 'crop',
+            resize: 'crop'
           },
           small: {
             w: 340,
             h: 226,
-            resize: 'fit',
+            resize: 'fit'
           },
           large: {
             w: 800,
             h: 532,
-            resize: 'fit',
-          },
-        },
-      },
-    ],
-  },
-});
+            resize: 'fit'
+          }
+        }
+      }
+    ]
+  };
+}